chore: Use text field in the region property
[#182299591](https://www.pivotaltracker.com/story/show/182299591)
