+++ google-stackdriver-trace.yml
@@ -18,8 +18,8 @@ id: bfb9c232-2b16-4d4c-8958-221ce1107406
 description: Beta - Distributed tracing service
 display_name: Stack driver trace (Beta)
 image_url: https://symbols.getvecta.com/stencil_4/52_stackdriver.23cab581ce.svg
-documentation_url: https://cloud.google.com/trace
-support_url: https://cloud.google.com/trace/docs/overview
+documentation_url: https://cloud.google.com/trace/docs
+support_url: https://cloud.google.com/support/
 tags: [gcp, tracing, stackdriver, beta]
 plans:
 - name: default
