+++ google-storage.yml
@@ -19,7 +19,7 @@ description: Beta - Google Cloud Storage that uses the Terraform back-end and gr
 display_name: Google Cloud Storage (Beta)
 image_url: https://cloud.google.com/_static/images/cloud/products/logos/svg/storage.svg
 documentation_url: https://cloud.google.com/storage/docs/overview
-support_url: https://cloud.google.com/storage/docs/getting-support
+support_url: https://cloud.google.com/support/
 tags: [gcp, storage, beta]
 plans:
 - name: private
@@ -64,30 +64,13 @@ provision:
         COLDLINE : COLDLINE
     - field_name: region
       type: string
-      details: The region of the mysql instance.
+      details: The region where the buckets are created.
       default: us-central1
-      enum:
-        asia-east1: asia-east1
-        asia-east2: asia-east2
-        asia-northeast1: asia-northeast1
-        asia-northeast2: asia-northeast2
-        asia-northeast3: asia-northeast3
-        asia-south1: asia-south1
-        asia-southeast1: asia-southeast1
-        australia-southeast1: australia-southeast1
-        europe-north1: europe-north1
-        europe-west1: europe-west1
-        europe-west2: europe-west2
-        europe-west3: europe-west3
-        europe-west4: europe-west4
-        europe-west6: europe-west6
-        northamerica-northeast1: northamerica-northeast1
-        southamerica-east1: southamerica-east1
-        us-central1: us-central1
-        us-east1: us-east1
-        us-east4: us-east4
-        us-west1: us-west1
-        us-west2: us-west2
+      constraints:
+        examples:
+          - us-central1
+          - asia-northeast1
+        pattern: ^[a-z][a-z0-9-]+$
     - field_name: credentials
       type: string
       details: GCP credentials
